+++ AntColonyCVRP/Cidade.h
@@ -3,17 +3,17 @@ class Cidade
 {
 public:
 	int id;
-	int posX;
-	int posY;
+	double posX;
+	double posY;
 	int demand;
 
 	Cidade(int id = 0, int posX = 0, int posY = 0, int demand = 0);
 
 	int getVertice() const;
 
-	int getPosX() const;
+	double getPosX() const;
 
-	int getPosY() const;
+	double getPosY() const;
 
 	int getDemand() const;
 };
